requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==2.8
importlib-metadata==7.0.1
ipython-genutils==0.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langchain-google-vertexai==0.0.1.post1
langsmith==0.0.82
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==3.0.0
platformdirs==4.1.0
portpicker==1.2.0
prompt-toolkit==1.0.18
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
shapely==2.0.2
simplegeneric==0.8.1
six==1.12.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.30.0
streamlit-chat==0.1.1
streamlit-chatbot==0.0.2.2
tenacity==8.2.3
terminado==0.13.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-protobuf==4.24.0.20240106
types-requests==2.31.0.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
zipp==3.17.0
